API/requirements.txt
Django==4.1.7
psycopg2>=2.8
djangorestframework>=3.9.0,<3.13.1
django-cors-headers==3.11.0
django-rest-auth==0.3.3
django-allauth==0.50.0
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.5
Pillow==9.4.0 
psycopg2==2.9.5
